requirements.txt
pygame-ce>=2.5.6
tomli>=2.0.0; python_version < "3.11"
